Platform/RaspberryPi/RPi3: Add reset and memory init libraries
The memory init library ensures that relevant memory regions are
reserved on boot.

The reset library supports the ResetSystem runtime call using PSCI
and signals the gRaspberryPiEventResetGuid event group on reset,
which we need to save modified configuration vars to NVRAM.

Note that we tried to drop ResetLib altogether, and use the reset
notification event provided by the EDK2, but this results in
Linux kernel panics, for which we have no workaround.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Jeremy Linton <jeremy.linton@arm.com>