Marvell/Armada7k8k: Remove SPCR baud rate setting
The mcbin (and likely others) have a nonstandard uart clock.
This means that the earlycon programming will incorrectly set
the baud rate if it is specified. The way around this is to tell
the kernel to continue using the preprogrammed baud rate. This
is done by setting the baud to 0.

Further, the SPCR and DSDT serial port need to match the port
address and port access type for the kernel to conclude they
are the same.

So while ARM_GAS32 is correct for earlycon (it can be used alone
on the kernel command line) by providing the reg-shift=2 value,
it also sets the io type to MMIO32, which doesn't match the DSDT
defined MMIO. This means that the actual console will never appear.
The obvious fix is to set reg-width=4 in DSDT, but that also changes
the accesssors to 32-bits (similarly to earlycon) and results in
console failure.

So the less obvious fix, is to use the GAS8 specifier. This means
that earlycon needs to be fully specified as
earlycon=uart,mmio32,0xf0512000, but has the extremely useful feature
that the console default works without any user interaction.

If in the future marvell decides to define their own ACPI id for the
console and upstream a quirk, the ARM_GAS8 portion of this should
be reverted.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>
Tested-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>