edk2-platforms: Change License.txt from 2-Clause BSD to BSD+Patent
Change License.txt in the root of the edk2-platforms repository from
the 2-Clause BSD License to the BSD+Patent License.

The text difference between these licenses in the patch is larger
than expected due to different choices for the position of the line
breaks.  The text from the BSD 2-Clause license and its disclaimer
are the same.  New clauses from the BSD+Patent license have been
added.

The License.txt files in other directories that contain the 2-Clause
BSD License are deleted.  The single License.txt in the root of the
edk2-platforms repository with the BSD+Patent License contents is the
preferred license for the edk2-platforms repository.

https://bugzilla.tianocore.org/show_bug.cgi?id=1373

This change is based on the following emails:

  https://lists.01.org/pipermail/edk2-devel/2019-February/036260.html
  https://lists.01.org/pipermail/edk2-devel/2018-October/030385.html

RFCs with detailed process for the license change:

  V3: https://lists.01.org/pipermail/edk2-devel/2019-March/038116.html
  V2: https://lists.01.org/pipermail/edk2-devel/2019-March/037669.html
  V1: https://lists.01.org/pipermail/edk2-devel/2019-March/037500.html

Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Thad Gillispie <thad.gillispie@intel.com>
Cc: Daocheng Bu <daocheng.bu@intel.com>
Cc: Isaac W Oram <isaac.w.oram@intel.com>
Cc: Maciej Piwko <maciej.piwko@intel.com>
Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Michael Kubacki <michael.a.kubacki@intel.com>
Cc: Shifei A Lu <shifei.a.lu@intel.com>
Cc: Xiaohu Zhou <bowen.zhou@intel.com>
Cc: Ankit Sinha <ankit.sinha@intel.com>
Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>