KabylakeSiliconPkg: Add FSP Dispatch switch.
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=1832

PcdFspModeSelection is used as switch for boot loader to
consume FSP in API mode or Dispatch mode. For backward
compatibility, if boot loader running in FSP API mode the
KabylakeFspBinPkg will be used which does not have FSP 2.1
changes. For FSP Dispatch mode AmberLakeFspBinPkg will be
used with FSP 2.1 support.
One INF was duplicate to include different DEC file in
each build type.

Test: Booted Kabylake RVP3 to Windows successfully.

Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Michael Kubacki <michael.a.kubacki@intel.com>
Signed-off-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>