Platform/Secure96Dxe: redefine LS connector CS as platform property
As opposed to the Secure96's I2C peripherals, whose bus addresses are
properties of the peripherals themselves, the SPI CS address of the
TPM is a property of the platform that incorporates the LS connector.

So tweak the macros that emit the CS values and related properties
to put it under the control of the platform that incorporates the
driver.

Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>