Vlv2TbltDevicePkg/PlatformInitPei: Update MemoryTypeInformation
Reduce extra reboots when UEFI Shell is run followed by an OS boot.
The extra reboot is due to use of EfiLoaderCode that exceeds the
MemoryTypeInformation bin size.  EfiLoaderCode should not be tracked.
Remove all memory types that should not be tracked, and update the
remaining bin sizes based on observed memory usage on an X64 DEBUG
profile.

Cc: Zailiang Sun <zailiang.sun@intel.com>
Cc: Yi Qian <yi.qian@intel.com>
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>
Reviewed-By: Zailiang Sun <zailiang.sun@intel.com>