Platform,Silicon RPi3: update license headers to BSD+Patent
RpiMbox.h and Bcm2836.h originate from Ard's original efforts, and were
always intended to be 2-clause BSD licensed, but due to a copy/paste
error, they ended up with a 3-clause BSD license instead - which meant
they were left out in Mike's major relicensing set.

(This is explained/confirmed in
https://edk2.groups.io/g/devel/message/40909
and
https://edk2.groups.io/g/devel/message/40911.)

This patch replaces the license headers in both with the SPDX statement
bringing it inline with the majority of the rest of the codebase.

Signed-off-by: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>