WhiskeylakeOpenBoardPkg: Add package and headers
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2083

Create the WhiskeylakeOpenBoardPkg to provide board support code. The
package may support Coffee Lake (CFL) and Whiskey Lake (WHL) boards. The
package serves as a board support package in the EDK II Minimum Platform
design. Silicon support for this package is provided in CoffeeLakeFspBinPkg
in the FSP repository and CoffeelakeSiliconPkg in the edk2-platforms
repository.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Ankit Sinha <ankit.sinha@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Ankit Sinha <ankit.sinha@intel.com>