Marvell/Drivers: MvSpiFlashDxe: Fix sector number obtaining
This commit fixes bug which was causing one sector of bytes to be
ommited. It was discovered when bytes to be written was 0 in last block,
but total count of bytes was multiplication of sector size. It turned
out that in every case one sector of data was missing.

This bug can be fixed in various ways, but this solution fixes hypotetical
situation in which total bytes count is smaller than sector size.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>