AdvancedFeaturePkg/Ipmi: Fix X64 GCC Build Failures
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2204

These build failures can be reproduced simply by building the
AdvancedFeaturePkg.dsc file in GCC5 for X64 architecture. To
build the whole package DSC (not pull individual features into
other packages), set the WORKSPACE variable to the edk2 directory
in the workspace as is done by executing edksetup.sh then create
the PACKAGES_PATH variable and add the Platform/Intel and
Silicon/Intel directories to the variable value. Then start the
build of AdvancedFeaturePkg.dsc:
  'build -p AdvancedFeaturePkg/AdvancedFeaturePkg.dsc -a X64'

This change adds EFIAPI to the following function declarations so
they are consistent with the declaration for module entry points.
  * BmcElog.c - InitializeBmcElogLayer ()
  * FrbPei.c - InitializeFrbPei ()
  * IpmiFru.c - InitializeFru ()

All future contributions to AdvancedFeaturePkg must successfully build
in GCC after this change.

Cc: Dandan Bi <dandan.bi@intel.com>
Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Dandan Bi <dandan.bi@intel.com>
Reviewed-by: Sai Chaganty <rangasai.v.chaganty@intel.com>