IntelSiliconPkg/FirmwareBootMediaLib: Add library
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2229

Adds a new library that is used to make system initialization
decisions in the boot flow dependent upon firmware boot media.
Note that the firmware boot media is the storage media that
the boot firmware is stored on. It is not the OS storage media
which may be stored upon a different non-volatile storage device.

Any Intel board, platform, or silicon code that must take action
conditionally based on the firmware boot media must use this
library API.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Sai Chaganty <rangasai.v.chaganty@intel.com>