CoffeelakeSiliconPkg: Add DSC include file section tags
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2246

The DSC LibraryClass files in CoffeelakeSiliconPkg that are intended
to be included elsewhere such as SiPkgPeiLib.dsc, SiPkgDxeLib.dsc,
and SiPkgCommonLib.dsc should have section tags so that they are
not dependent on the top-level DSC file to place the include file
in the correct location in the DSC file and better define the
applicability of their library content.

This change adds section tags for the library class related files.
The component files may be built differently in the consuming package
depending on their architecture requirements so those are not
modified.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Chasel Chiu <chasel.chiu@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>