KabylakeOpenBoardPkg: Assign unique token namespace
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2247

PCDs declared in the KabylakeOpenBoardPkg currently use the GUID
gBoardModuleTokenSpaceGuid. The same name is used in other board
packages and a package has been added called BoardModulePkg so
this name is now misleading.

This change assigns a unique GUID value and a name specific to the
package to provide differentiation from PCDs in other board
packages.

Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Jeremy Soller <jeremy@system76.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>