SimicsOpenBoardPkg: Use new advanced feature refactor
Updates advanced feature references in SimicsOpenBoardPkg to the
new advanced feature packages in Features/Intel. Previously, the
advanced features in Platform/Intel were used.

SimicsOpenBoardPkg currently uses the SMBIOS and Network advanced
features. These are the same features used by the package prior to
the advanced feature refactor.

Note that SimicsOpenBoardPkg depends upon these advanced features
even when the boot stage is less than the advanced feature stage.

This bug was not fixed in this commit and is filed to be resolved
in the following Bugzilla item:
  https://bugzilla.tianocore.org/show_bug.cgi?id=2165

Cc: Wei David Y <david.y.wei@intel.com>
Cc: Agyeman Prince <prince.agyeman@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>