Platform/Intel: Add FIT generation tool
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2210
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2211

Added FitGen tool build and FIT generation to the BIOS
build process.

What was done:

- Build the FIT generation tool
- Added default/empty BIOS_INFO_GUID to the build.cfg
- Added BIOS_INFO_GUID to GalagoPro3,KabylakeRvp3 and
  WhiskeylakeURvp's build_config.cfg

This allows a board to specify the GUID associated with
the BIOS Info PEIM to be used in the board's FIT generation.

BIOS_INFO_GUID is passed as an argument to FitGen tool
which allows the tool to locate the BIOS Info module to be
used in FIT generation.

Cc: Ankit Sinha <ankit.sinha@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Michael Kubacki <michael.a.kubacki@intel.com>

Signed-off-by: Prince Agyeman <prince.agyeman@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>
Reviewed-by: Michael Kubacki <michael.a.kubacki@intel.com>