SimicsOpenBoardPkg/BoardX58Ich10: Use PEI_ARCH and DXE_ARCH
This change updates OpenBoardPkg.dsc to use the PEI_ARCH and
DXE_ARCH specified in the DSC file for the board package components.

An EDK II build limitation currently prevents the macros from
being able to be used, so a todo has been placed with the action
needed when the BaseTools update is completed as described in the
following Bugzilla item:
  https://bugzilla.tianocore.org/show_bug.cgi?id=2308

The intention of this macro is to ensure that all DSC files that
compose the package build (such as those included from
MinPlatformPkg or an advanced feature package) build phase-specific
drivers for the architecture defined by the board package.

This is a macro that is required in all Minimum Platform board
packages.

Cc: Wei David Y <david.y.wei@intel.com>
Cc: Agyeman Prince <prince.agyeman@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>