Silicon/Broadcom: Add Bcm2711 header
Bcm2711 is the SoC used by the Raspberry Pi 4. For the most part it
is backward compatible with the Bcm283x SoC used by older models but
some new Bcm2711 registers have been introduced such as the ones for
PCIe.

We therefore introduce a new Bcm27xx module, which contains only the
base address for the new registers for now (set from a PCD as is the
case for Bcm283x) but which should get populated further as we add
support for the Raspberry Pi 4.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>
Reviewed-by: Philippe Mathieu-Daude <philmd@redhat.com>