Features/Intel/AdvancedFeaturePkg: Add package
This change adds AdvancedFeaturePkg to the Features/Intel
directory. AdvancedFeaturePkg is a special package that is
located at the root level of Features/Intel.

All advanced features can be built using the AdvancedFeaturePkg
build. To build all advanced features, follow the instructions in
the Feature/Intel/Readme.md file. The actual build is performed by
executing the following command in the Features/Intel directory:

"build -p AdvancedFeaturePkg/AdvancedFeaturePkg.dsc -a IA32 -a X64"

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>