KabylakeSiliconPkg: Remove redundant ResetSystem() implementation
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=2375

The following commits introduced a new implementation of
ResetSystem () in ResetLib instances in KabylakeSiliconPkg:
 * 2393cff0d4
 * 07fd95c9ca
 * 0ad2f057d8
 * a6e61d43c9

The PchResetLib instance PeiPchResetLib links against ResetLib and
already contained a definition of ResetSystem (). This resulted in
a build failure in VS2017 due to multiple definitions of a symbol
being provided to the linker. The failure did not reproduce in
VS2015 so it persisted until now.

This change removes the pre-existing implementation of
ResetSystem () in PchReset.c as it is now linked by the
implementation in the ResetLib instance.

Cc: Sai Chaganty <rangasai.v.chaganty@intel.com>
Cc: Chasel Chiu <chasel.chiu@intel.com>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>
Cc: Zhichao Gao <zhichao.gao@intel.com>
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>
Reviewed-by: Chasel Chiu <chasel.chiu@intel.com>
Reviewed-by: Nate DeSimone <nathaniel.l.desimone@intel.com>