Platform/DeveloperBox: disable variable runtime cache
The variable runtime cache for SMM enabled implementations of the
variable runtime DXE driver was introduced after the standalone
MM based implementation was merged for the DeveloperBox platform.

This means the combined binary image of ARM Trusted Firmware and
the standalone MM runtime we carry in edk2-non-osi predates this
feature, so we need to disable it when building the non-secure side
when incorporating the prebuilt binary.

Whether it makes sense to enable the runtime cache for this platform
is a different question, so let's keep it disabled entirely until
someone identifies a need for it. Note that this patch is against the
include file that is shared between the standalone MM and the DXE
sides of the varstore implementation, so things should work as
expected when building both sides from scratch.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>