Silicon/Broadcom/Bcm283x: Rename Bcm2835 RNG driver
The Bcm283x architecture provides multiple implementations of the
Random Number Generator, that have incompatibilities with one
another.

Because we will need to add a new driver for the Bcm2838 RNG we
rename the existing Bcm2835-compatible driver to Bcm2835RngDxe
and update the one platform that depends on this driver (RPi3).

We also rename the internal function calls from Bcm2836 to
Bcm2835 since, if we ever add platform support for Pi models
that use this SoC (original Pi, Pi Zero, etc), it may look
confusing to reference a feature that seems to belongs to the
next SoC iteration.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Philippe Mathieu-Daude <philmd@redhat.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>