Silicon/Bcm27xx: Add segment library to handle nonstandard ECAM
This SoC uses a nonstantard ECAM with requires the provision of a custom
segment library. Basically, with the Bcm2711, the root port is the first
bytes of the register space (offset 0) and individual devices are selected
by computing their BDF index and writing that into the CFG_INDEX register
before ECAM data can then be read/written at CFG_DATA.

We also ensure that read/write accesses are serialized through the use of
a lock, as some of the library calls cannot run concurrently.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>