Platform/RPi/SmbiosDxe: Report a more human readable firmware revision
The firmware revision that is queried through the VideoCore mailbox
is really the 32-bit timestamp of when the VideoCore firmware was
generated.

To make this more palatable for human reporting, convert it to a
YY.MM firmware revision, so that end-users can get an approximative
idea of how old their VideoCore firmware is.

Signed-off-by: Pete Batard <pete@akeo.ie>
Reviewed-by: Philippe Mathieu-Daude <philmd@redhat.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>