Silicon/SynQuacer/NetsecDxe: ignore DMA coherency at binding time
Whether the Netsec driver is built to support coherent or non-coherent
DMA is decided by the DmaLib library class resolution that is provided
by the platform description, and there are various other places where
the platform needs to be consistent with this (ACPI tables, device
tree, platform driver, etc)

Since the driver has no way to figure out which flavor of DmaLib it was
built against, there is no point in limiting it to one flavor only, and
we can drop the DmaType check at binding time entirely.

This fixes an issue introduced by commit 88c7874481b856578662b775
("Silicon/SynQuacer/PlatformDxe: set NETSEC DMA as coherent"), which
incorrectly stated that the driver does not look at the DmaType flag
in the first place.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>