Silicon/Intel/Tools: Add parameter for microcode alignment in FitGen.
The current FitGen has "-NA" parameter to indicate whether microcode
is placed with an alignment, but it could only support 0x800 alignment:
 - With"-NA" means microcode is not aligned.
 - No "-NA" means Microcode is 0x800 aligned.
There is no method to specify other alignment value.

This patch add "-A" option to FitGen for to configure the alignment
to a user specified value. The change is backward compatible as:
 - Only "-NA" means microcode is not aligned (same as before).
 - No "-NA" and No "-A" means Microcode is 0x800 aligned (same as
   before).
 - Only "-A" means microcode is aligned with specified value (new).

Cc: Bob Feng <bob.c.feng@intel.com>
Cc: Liming Gao <liming.gao@intel.com>
Signed-off-by: Siyuan Fu <siyuan.fu@intel.com>
Reviewed-by: Bob Feng <bob.c.feng@intel.com>